Parameterize PSS base URL
Replace various hardcoded values for the Primary Source Sets API URL
with a new constant. Modify various files in primary-source-sets,
introducing the PSS_BASE_URL constant.

Additionally:

* In pages/primary-source-sets/set/teaching-guide.js, fix the URL of the
AJAX endpoint that's used to get information on the set. The value is
taken from the JSON response of an earlier AJAX request, which returns
the _page_ of the teaching guide. We serve the PSS API from a location
that's configured by the variable PSS_BASE_URL, so the value for the
page needs to be processed to make sure it's pointing at the correct API
endpoint.

* Fix missing key warning (mgiraldo)