Use all available CPUs (#765)
Use the Node.js standard library "cluster" module to fork as many
workers as there are CPUs, for better performance.

The `.catch` event handler that used to be applied to `app` has been
removed.  I do not believe that it was reachable, as I have tried
inserting `throw` statements in various places in the `then` closure,
only to find them not handled by the `catch` handler. Next.js apparently
handles them.

The `process.on("uncaughtException"...)` handler, meanwhile, does catch
exceptions that occur outside of `app.prepare().then()`, but its job is
to end execution after printing the error. This seems to be what the
original author thought the `app.prepare().then().catch()` was going to
do.