Adds new application api endpoint for aboutness (#968)
Initial implementation of aboutness (links under `Explore related items in DPLA` on this page: http://recollectionwisconsin.dp.la/search?q=food) was returning unexpected results and did filtering of items on the client side.

Short version of how feature should work:
`It should show things that are subject=Wisconsin OR location Wisconsin and are not provider = Wisconsin, and the search term should be what the search term is for the page`

I built a new internal API endpoint that does this via a Lucene syntax query and pointed the aboutness links at that instead. Removed a bunch of the client side filtering code as a result.