slay seadragon (#1126)
This uses a Cypress api to get access to the window object and trims out the image viewer on exhibitions pages that have images.

The openseadragon viewer has a mouseover control overlay that fades out using an opacity attribute and a js timeout, and it's unpredictable where that process is going to be when Cypress goes to do a snapshot. So I'm just removing it from the DOM for snapshots until I understand better how to deal with that in cypress. Ideally I'd just remove the overlay, but I haven't figured out the selectors yet.